feat(agent): Process context and reorganized storage reconcilers
