+++ crates/agent/src/reconcile/agent.rs
@@ -1,27 +1,21 @@
-use std::{
-    collections::HashSet,
-    path::Path,
-    sync::Arc,
-    time::{Duration, Instant},
-};
+use std::{collections::HashSet, sync::Arc, time::Instant};
 
 use futures::stream::AbortHandle;
 use snops_common::{
     api::EnvInfo,
-    binaries::{BinaryEntry, BinarySource},
-    constant::{SNARKOS_FILE, VERSION_FILE},
+    binaries::BinaryEntry,
     rpc::error::ReconcileError2,
-    state::{
-        AgentId, AgentPeer, AgentState, InternedId, NetworkId, NodeState, StorageId, TransferId,
-    },
+    state::{AgentId, AgentPeer, AgentState, NetworkId, NodeState, StorageId, TransferId},
 };
 use tarpc::context;
-use tokio::sync::{Mutex, Semaphore};
-use tracing::{error, trace, warn};
+use tokio::sync::Mutex;
+use tracing::{error, warn};
 
 use super::{
-    command::NodeCommand, default_binary, get_version_from_path, DirectoryReconciler,
-    FileReconciler, Reconcile, ReconcileCondition, ReconcileStatus,
+    command::NodeCommand,
+    process::ProcessContext,
+    storage::{BinaryReconciler, StorageVersionReconciler},
+    DirectoryReconciler, Reconcile, ReconcileStatus,
 };
 use crate::state::GlobalState;
 
@@ -73,12 +67,12 @@ impl TransfersContext {
 
 #[derive(Default)]
 pub struct AgentStateReconcilerContext {
-    /// All parameters needed to build the command to start the node
-    command: Option<NodeCommand>,
-    /// Information about active transfers
-    transfers: Option<TransfersContext>,
     // TODO: allow transfers to be interrupted. potentially allow them to be resumed by using the
     // file range feature.
+    /// Information about active transfers
+    transfers: Option<TransfersContext>,
+    /// Information about the node process
+    process: Option<ProcessContext>,
 }
 
 impl Reconcile<(), ReconcileError2> for AgentStateReconciler {
@@ -104,6 +98,10 @@ impl Reconcile<(), ReconcileError2> for AgentStateReconciler {
                 // If the node should be torn down, or the storage has changed, we need to
                 // gracefully shut down the node.
                 let shutdown_pending = !node.online || storage_has_changed;
+
+                if let (true, Some(process)) = (shutdown_pending, self.context.process.as_ref()) {
+                    // TODO: reconcile process destruction
+                }
 
                 // TODO: check if addrs have changed, and update shutdown_pending
 
@@ -189,7 +187,7 @@ impl Reconcile<(), ReconcileError2> for AgentStateReconciler {
                 )
                 .await?;
 
-                if self.context.command.as_ref() != Some(&command) {
+                if self.context.process.as_ref().map(|p| &p.command) != Some(&command) {
                     // TODO: OK to restart the node -- command has changed
                 }
 
@@ -281,123 +279,6 @@ impl Reconcile<(), ReconcileError2> for AddressResolveReconciler {
         }
 
         Ok(ReconcileStatus::default())
-    }
-}
-
-/// Download a specific binary file needed to run the node
-struct BinaryReconciler<'a> {
-    state: Arc<GlobalState>,
-    env_info: Arc<EnvInfo>,
-    node_binary: Option<InternedId>,
-    /// Metadata about an active binary transfer
-    binary_transfer: &'a mut Option<(TransferId, BinaryEntry)>,
-    /// Time the binary was marked as OK
-    binary_ok_at: &'a mut Option<Instant>,
-}
-
-impl<'a> Reconcile<(), ReconcileError2> for BinaryReconciler<'a> {
-    async fn reconcile(&mut self) -> Result<ReconcileStatus<()>, ReconcileError2> {
-        let BinaryReconciler {
-            state,
-            env_info,
-            node_binary,
-            binary_transfer,
-            binary_ok_at,
-        } = self;
-
-        // Binary entry for the node
-        let default_binary = default_binary(env_info);
-        let target_binary = env_info
-            .storage
-            .binaries
-            .get(&node_binary.unwrap_or_default())
-            .unwrap_or(&default_binary);
-
-        // Check if the binary has changed
-        let binary_has_changed = binary_transfer
-            .as_ref()
-            .map(|(_, b)| b != target_binary)
-            .unwrap_or(true);
-        let binary_is_ok = binary_ok_at
-            .map(|ok| ok.elapsed().as_secs() < 300) // check if the binary has been OK for 5 minutes
-            .unwrap_or(false);
-
-        // If the binary has not changed and has not expired, we can skip the binary
-        // reconciler
-        if !binary_has_changed && binary_is_ok {
-            return Ok(ReconcileStatus::default());
-        }
-
-        let src = match &target_binary.source {
-            BinarySource::Url(url) => url.clone(),
-            BinarySource::Path(path) => {
-                let url = format!("{}{}", &state.endpoint, path.display());
-                url.parse::<reqwest::Url>()
-                    .map_err(|e| ReconcileError2::UrlParseError(url, e.to_string()))?
-            }
-        };
-        let dst = state.cli.path.join(SNARKOS_FILE);
-
-        let is_api_offline = state.client.read().await.is_none();
-
-        let file_res = FileReconciler::new(Arc::clone(state), src, dst)
-            .with_offline(target_binary.is_api_file() && is_api_offline)
-            .with_binary(target_binary)
-            .with_tx_id(binary_transfer.as_ref().map(|(tx, _)| *tx))
-            .reconcile()
-            .await?;
-
-        // transfer is pending or a failure occurred
-        if file_res.is_requeue() {
-            return Ok(file_res.emptied().add_scope("file_reconcile/requeue"));
-        }
-
-        match file_res.inner {
-            // If the binary is OK, update the context
-            Some(true) => {
-                **binary_ok_at = Some(Instant::now());
-                Ok(ReconcileStatus::default())
-            }
-            // If the binary is not OK, we will wait for the endpoint to come back
-            // online...
-            Some(false) => {
-                trace!("binary is not OK, waiting for the endpoint to come back online...");
-                Ok(ReconcileStatus::empty()
-                    .add_condition(ReconcileCondition::PendingConnection)
-                    .add_scope("agent_state/binary/offline")
-                    .requeue_after(Duration::from_secs(5)))
-            }
-            None => unreachable!("file reconciler returns a result when not requeued"),
-        }
-    }
-}
-
-struct StorageVersionReconciler<'a>(&'a Path, u16);
-
-impl<'a> Reconcile<(), ReconcileError2> for StorageVersionReconciler<'a> {
-    async fn reconcile(&mut self) -> Result<ReconcileStatus<()>, ReconcileError2> {
-        let StorageVersionReconciler(path, version) = self;
-
-        let version_file = path.join(VERSION_FILE);
-
-        let version_file_data = if !version_file.exists() {
-            None
-        } else {
-            tokio::fs::read_to_string(&version_file)
-                .await
-                .map_err(|e| ReconcileError2::FileReadError(version_file.clone(), e.to_string()))?
-                .parse()
-                .ok()
-        };
-
-        // wipe old storage when the version changes
-        Ok(if version_file_data != Some(*version) && path.exists() {
-            let _ = tokio::fs::remove_dir_all(&path).await;
-            ReconcileStatus::default()
-        } else {
-            // return an empty status if the version is the same
-            ReconcileStatus::empty()
-        })
     }
 }
